Richer info in contacts search suggestions.
Before, the search suggestions returned by the ContactsProvider only
showed a single line, with the contact name, organization, phone number
or e-mail. Global search had it's own version of contacts suggestions,
adding contact photos, a second text line, and IM presence.

This patch moves the richer information to the suggestions returned
by Contactsprovider. Advantages of this:
- The search suggestions in the Contacts app will now be richer.
- In global search, contacts suggestions will now look the same
when selecting the Contacts source as they do in the blended
suggestion list.
- The contacts-specific code can be removed from GlobalSearch.