Add code for setting a new primary data record.
Adds private methods setIsPrimary() and setIsSuperPrimary(). When a
client of the content provider attempts to update the "is_primary" or
"is_super_primary" columns on a data record, setIsPrimary() or
setIsSuperPrimary() is called.

setIsPrimary() will set the "is_primary" field on the target data
record to 1, and set all other data records of the same mimetype
associated with the contact to 0.

setIsSuperPrimary() will set the "is_super_primary" field on the target
data record to 1, and set all other data records of the same mimetype
associated with the aggregate to 0.

Updating the "is_primary" or "is_super_primary" column to zero explicitly
by a client will be ignored.