Implemented core of restricted data security model.
The provider now supports the notion of a Contact being
flagged with IS_RESTRICTED, meaning that only packages with
granted RestrictionExceptions can read or write the
data contained in that Contact.  The owning PACKAGE of a
Contact can dynamically grant and revoke exceptions to
allow itself and other packages to read the contained data.

RestrictionExceptions are granted through update() calls
to ContactsProvider, and we enforce that the calling process
actually has authority over the package it's changing.  The
exceptions are stored to both a database table and into an
in-memory lookup table.

Created applyDataRestrictionExceptions() helper method that
enforces IS_RESTRICTED flags while also adding clauses to
handle any RestrictionExceptions granted to the calling
process.  A similar approach is used for aggregate summary
queries, where we return the "best" accessible super-primary
phone number available to the caller.  Internally we keep
track of an "optimal" primary that may be restricted, and a
"fallback" primary that is always readable.

Wrote a @Large test suite to exercise the permissions
granting and revoking infrastructure, and to test the
internal super-primary storage.