Fix column map naming, empty projection behavior.
The "MAX(mode)" projection needed to be named as the
expected column to be useful.  The default behavior of
assertContained when asking for no specific columns
(projection=null) should be to /not/ limit to just that
single column.  And finally, fix up one place where the
protection clause was being appended wrong.