Fixing breakage of legacy contacts API support broken by an earlier CL
The gist of the issue is that SQLiteDatabase supports one nested transaction, but only one.
The code was executing two nested transactions (not two levels of nesting, just two sequential
transactions) and that was causing the outer transaction to roll back.  We really should fix
this in SQLiteDatabase.

So the provider will now call the aggregator before, not after transaction commit, and
the aggregator will not open a nested transaction.

Change-Id: I9fa09d36ea43a62ba4f2fd5e5e1816cc439d225b