Making contact name filtering more restrictive: ignoring name variants and email-based nicknames.
I discovered this issue while trying to reproduce http://b/2123390
I typed "h" and saw Erica among the filtered contacts.  The reason is that the nickname cluster
id for Erica matched the hex representation of letter h.  The fix is to be more specific about
what types of name lookup we are interested in.

Change-Id: If367424052ee8e93757c98ce1dbb435522d8ba2e