Offer to delay visibility updates until caller forces.
We now watch for query parameters on insert, update, and
delete for Settings and Groups, which we use to delay
updating of IN_VISIBLE_GROUP at the callers request, which
is going to be used to help avoid ANRs.  Also wrote unit
test to verify it works.  Part of fixing http://b/2075275