Change selection argument so that it complies with the current view given from ContactsProvider2.
Root cause: VCardComposer changed the table to look over, from
RawContacts to Contacts, which is with valid ratinale
(Contacts shows aggregated view, while RawContacts does not).

Unfortunately, the change made ContactsProvider2 corrupted, which has
assumed the table VCardComposer is looking at is RawContacts.

Now, after discussing with Contacts team, this change is concluded to
"correct" change.

I changed from RawContacts._ID to Contacts._ID after jsharkey's approval,
but the actual string is the same "_id", so the essential is the same.
The difference is that readers would not be confused with seeing "RawContacts"
in the wrong context.

Internal issue id: 2162658