Optimizing the email filter query.
In order to make the query perform, I had
to break out two separate nested queries:
one for the email lookup and one for name lookup,
and build the UNION of found _ids before
retrieving suggestions from the Data table.

Performance of the query is improved by orders of
magnitude.

All tests pass.

Bug: http://b/2286776

Change-Id: I0ff20aa2bd3bcfc765831390a36274c01e0e16e3