Removing unnecessary limit on name lookups in global search support for contacts
Having a limit on the number of name lookups was a bad idea
to begin with: you may be looking for one visible contact and
name lookup would yield 500 matches of which 499 are invisible,
so you would consistenly miss the one contact you were looking
for.

I have confirmed that removing this limit does not affect
the performance of the query too severely. With a set of 5000
contacts all starting with the letter "D" I still got the
one contact I was looking for in less than a second on Sholes.


See bug http://b/2298674

Change-Id: Id4219dea1be2f9ac885ef5ddda5e8376c3e050ee