Add GoldenContact, SuggestionTester and their builders.
Use builders in place of an assert method with many arguments, so it's
easier to see which argument has which value.  In addition, separate the
creation of the contact from the testing of the global search.  Hopefully
the builder for contact creation will be useful for other tests as well.

Change-Id: Idbc8571a37b9b116198f1e00da6b2ecd5a22235f