Optimizing the Contacts Display Setting screen.
Summary:
Given an account with 5020 contacts,
Before this optimization it would take roughly 10sec to
change visibility of either 5000 of them or 20 of them.
After the optimization, changing visibility of 5000 contact
still takes roughly the same time, but changing visiblity
of 20 contacts takes 1.5 sec.

There is another optimization that can be done (mentioned in the
CL that reduces the cost for 5000 contacts from 10sec to
about 3.5sec.  That's a bit more involved and a TODO for later.


Change-Id: Iee8f251fdcab50a26b6b02131c0d23fb84aabd46