Fixing phone number picker in search mode.
The phone number picker crashes if you
try to search by something that is neither
a name nor a phone number, e.g. "&(./"

This is a relatively rare use case for people,
but very common for monkeys. The main reason to
fix it now is to allow monkeys to continue running
and discovering other problems.

Bug: 2473955
Change-Id: Ib48e324b4d3ced262cc90a67471a35765742eff1