Start weaning NameNormalizer off icu4jni implementation details.
Minor reshuffling necessary to keep building with the Java 6 API changes
post-froyo. This is compatible with the Java 5 API, and needs only an import
change to be compatible with the Java 6 API.

Ideally, this class should only be using public API, but I don't think
anyone has time for that right now.

Bug: 2417080
Change-Id: I40ac70a62ccf0a096a5635cc0499699b8bb69498