Unit tests for voicemail provider.
These tests cover basic functionality of the provider including
permission checks and media content input/output.

The key functionality that is yet to be tested is provider
change broadcast intents. This requires us to use a mocking framework,
and we are yet to finalize on which one we will use.

Change-Id: I2304309c4fc109cc1e0b969ede33d8268a4d4194