Fixed VoicemailProviderTest.
A recent change in voicemail content provider used the package manager
queryBroadcastReceivers() to query the receiving packages of the broadcast intent.
However, mock package manager used in the test env does not support this
operation and throws an exception.
The fix is to wrap this method within the content provider
implementation and stub the wrapper method in the test.

Change-Id: Ief4887f1379614466ea6471cf7b132ef873550ba