Do not send the provider_changed event to pkgs that have not changed.
The earlier code sent provider changed notification with permission
READ_WRITE_ALL_VOICEMAIL if the candidate package is not affected by the
change. This guaranteed that the change is not notified to the package
if it does not have ALL permission, but it generated an error log with
"permission denied" message. This is not desirable.

This change checks for ALL permission explicitly and do not even fire
the notification if the candidate package is not affected and does not
have ALL permission.

Change-Id: I20982572d7f9a08ce98b03b23c9d2d8aa60efc9f