Create account record if needed on profile insert.
For accounts without a data set, this should be handled automatically
when the account is added.  But with data sets, we need to handle it
specially so that the profile raw contact ID reference has a row to
hang off of.

Other raw contacts or groups with a data set will have the account
record created in the background when the provider is initialized.

Change-Id: I24ea0c389c3bc839aee295f1297e2236c473d66f