Fix photo updates when no photo is specified.
If no photo key is present, we will not attempt to do any photo
processing on the row.  The existing photo and photo file, if any,
will be unaffected.

If a photo key is present but the value is null or an empty byte
array, we will clear out the photo BLOB and photo file fields when
updating.

Bug 5247668

Change-Id: Ice8508060ab272bedfb42d105df55e972357e843