Fix test crash in ContactsProvider2.initialize().
Seems like Secure.getLong() (and other methods) internally tries to get
the Settings provider from the passed context when it's called first time,
and it crases in tests because our mock context doesn't have it.

Let's just call it with an actual contxext when tests start so it'll initialize
the settings provider correctly.

Bug 5499268

Change-Id: I863862ead2078715621cb71b1cf15f0dd94c0feb