Cherry-pick Fix for GB upgrade issue with phone lookups.
This is a cherrypick from master for this CL
Iea81bc4163cb89169946f9a33f9e15b13e8868ad

This just generalizes a fix that Shaopeng had for Brazilian phone
numbers, in which it was possible that the caller ID number being
checked was shorter than the normalized number we had stored.  We
essentially accept a phone number as a match if the shorter of
either (caller ID, normalized number) is a trailing prefix of the
other.

Bug: 5638376
Change-Id: Id61d5295272106a05af14d3c2ce1695eaf2dd372