Clean up Phone.NORMALIZED_NUMBERS upgraded from GB
If upgraded from GB, the NORMALIZED_NUMBERS column doesn't really
contain the "normalized" numbers.  It's some kind of stripped-out-and-reversed
numbers.

We can't create normalized numbers for them as we don't really know the correct
country for each number (and we can't even rely on CountryDetector during the
upgrade step -- the detector isn't ready at this moment), so let's just null
out those bad numbers.  This is how we do for invalid numbers.

phone_lookup.NORMALIZED_NUMBERS also has the same issue, so let's recreate
them in the way DataRowHandlerForPhoneNumber.insert() would when it can't
infar the normalized number.  (i.e. when the phone number is invalid.)

Bug 5690851

Change-Id: I756b828e15a50a2f711c01ef22abd2cfa67bad0a