Add debug activity to export all data files as a zip
This will allow us to collect contacts database files even from user build
devices where "adb root" is disabled.

This is simialr to what CalendarProvider does in CalendarDebugActivity.  The
difference is it'll export all files under "/data/data/com.android.providers.contacts/",
including the profile db and highres photo files.

To launch the activity:
adb shell am start -a com.android.providers.contacts.DUMP_DATABASE

I'm planning to add somethig to the people app to fire off this intent, so that
users will be able to do it without adb.

The activity will show a warning message with "Start", "Delete" and "Cancel"
buttons.

- "Cancel" will close the dialog.

- "Start" will create a zip file.  After that, it'll present the activity
chooser to let the user choose which app to use to send it with.

- "Delete" will delete the ZIP file.  We need to make sure to ask the user to
do this once sending email is succeessfully finished (unfortunately there's
no way for us to detect it programmatically), as any apps with the "read
sdcard" parmission will be able to read it otherwise.

In the future, we may want to add options to, for example, exclude the profile
db or exclude highres pictures.

Bug 6813842

Change-Id: Id181efad65194ed39b0a0bc1226252da62b8927e