Make "export contacts database" more secure
Don't put the dump file on the SD card.  Instead, put it in the internal cache
directory which is protected by the filesystem permissions.

In order to make it attachable on gmail, create a shim content provider and
sends a content: URI for this provider.

The dump file can be read only from the apps that knows its name, which
we pass via the SEND intent.  Each dump file has a unique 256-bit random name,
so it's virtually impossible to for other apps to read them.

Bug 6813842

Change-Id: I3ca081e696e4e432e2bf7eb701595c508cd19409